fix: address review feedback (remove base class/trait for Vonage)
